# docs/apispecs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kasir App",
        "description": "Kasir App API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/v1"
        }
    ],
    "paths": {
...
        "/user/register": {
            "post": {
                "tags": [
                    "User API"
                ],
                "description": "Register User",
                "summary": "Register User",
                "operationId": "registerUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/user_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/category/{category_id}": {
            "get": {
                "tags": [
                    "Category API"
                ],
                "description": "Get Category",
                "summary": "Get Category",
                "operationId": "getCategory",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/categories_response"
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category API"
                ],
                "description": "Delete Category",
                "summary": "Delete Category",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "message": {
                                                        "description": "category deleted successfully",
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category": {
            "post": {
                "tags": [
                    "Category API"
                ],
                "description": "Create Category",
                "summary": "Create Category",
                "operationId": "createCategory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Category Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Category API"
                ],
                "description": "Get Categories",
                "summary": "Get Categories",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Categories Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Product API"
                ],
                "description": "Get Products",
                "summary": "Get Products",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Products Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "product_type": {
                                                        "type": "string"
                                                    },
                                                    "stock": {
                                                        "type": "number"
                                                    },
                                                    "code_product": {
                                                        "type": "string"
                                                    },
                                                    "base_price": {
                                                        "type": "number"
                                                    },
                                                    "selling_price": {
                                                        "type": "number"
                                                    },
                                                    "category_id": {
                                                        "type": "number"
                                                    },
                                                    "minimum_stock": {
                                                        "type": "number"
                                                    },
                                                    "shelf": {
                                                        "type": "string"
                                                    },
                                                    "weight": {
                                                        "type": "number"
                                                    },
                                                    "discount": {
                                                        "type": "number"
                                                    },
                                                    "information": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product API"
                ],
                "description": "Create Product",
                "summary": "Create Product",
                "operationId": "createProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "product_type": {
                                        "type": "string"
                                    },
                                    "stock": {
                                        "type": "number"
                                    },
                                    "code_product": {
                                        "type": "string"
                                    },
                                    "base_price": {
                                        "type": "number"
                                    },
                                    "selling_price": {
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    },
                                    "minimum_stock": {
                                        "type": "number"
                                    },
                                    "shelf": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "number"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "information": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/products_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/customers": {
            "post": {
                "tags": [
                    "Customers API"
                ],
                "description": "Create User",
                "summary": "Create User",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "full_name": {
                                                        "type": "string"
                                                    },
                                                    "phone": {
                                                        "type": "number"
                                                    },
                                                    "address": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "get": {
                "tags": [
                    "Customers API"
                ],
                "description": "Get User",
                "summary": "Get User",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/customers_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
            "schemas": {
...
                "user_response": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
                "categories_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "name": {
                            "type": "string"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
                "products_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "name": {
                            "type": "string"
                        },
                        "product_type": {
                            "type": "string"
                        },
                        "stock": {
                            "type": "number"
                        },
                        "code_product": {
                            "type": "number"
                        },
                        "base_price": {
                            "type": "number"
                        },
                        "selling_price": {
                            "type": "number"
                        },
                        "category_id": {
                            "type": "number"
                        },
                        "minimum_stock": {
                            "type": "number"
                        },
                        "shelf": {
                            "type": "string"
                        },
                        "image": {
                            "type": "string"
                        },
                        "weight": {
                            "type": "number"
                        },
                        "discount": {
                            "type": "number"
                        },
                        "information": {
                            "type": "string"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
                "customers_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "full_name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
            },
            "securitySchemes": {
                    "BearerAuth": {
                        "type": "http",
                        "scheme": "bearer",
                        "bearerFormat": "JWT"
                    }
                }
        }
}